Big QOL Update and Fixes
Added CPUInfo, to query cpuid features
Query RAM information
Query Windows version
Made IsWithin/IsWithinExclusive/IsWithinInclusive more consistent
Improved StringUtils with support for StringViews
Fixed StringUtils::Trim
Minimum Windows supported version is 8
Added Reflection to Math types
Added ToString/FromString to all math types
TestApplication will containt a manifest that allows it to use Vista/Win7/Win8/Win8.1/Win10 features

+++ src/Public/Math/Base/Utils.inl
@@ -279,7 +279,7 @@ namespace greaper::math
 	{
 		return std::atan2f(y, x);
 	}
-	INLINE float ATan2(double y, double x)noexcept
+	INLINE double ATan2(double y, double x)noexcept
 	{
 		return std::atan2(y, x);
 	}
